api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
httpx==0.28.1
qdrant-client==1.11.3
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
PyJWT==2.9.0
pytest==8.3.5
